wip: clean up the core crate after switching to recursive diff engine
